requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
django-cors-headers==3.6.0
django-filter==2.4.0
django-money==1.2.1
django-oauth-toolkit==1.3.3
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-social-oauth2==1.0.8
freeze==3.0
gunicorn==20.0.4
idna==2.10
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
Pillow==8.0.1
psycopg2-binary==2.8.6
py-moneyed==0.8.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.2.0
wrapt==1.12.1
